Ampere: EthernetInterface: Reimplement change IPv4Addresses property.
Users should send requests with data in the same order as data got from
GET response. The server will ONLY change IPv4 records which have non-empty
slot from the requests.

Because the old approach doesn't restart the networkd to apply it into
the system. The new way does the following steps:
  * Delete the old IPv4 record which needs to be replaced.
  * Create new IPv4 record with new data.

Caveat: If DHCP is enabled, and changed IPv4 data equals to the old dynamic
address, it will delete the old one, and no IPv4 is created. This
behavior is same as create IPv4.

This commit also does some clean-ups and remove the redundant check of
addressOrigin because createIPv4 doesn't accept it.

Tested under real hardware. No regression and working.

Implements #2138.

Change-Id: I19955189b5b5001d8c1cf0eedfad1f845a7117b7
Signed-off-by: Hy Che <hche@amperecomputing.com>