Ampere: ethernet: Initialize IPv4 values before loading...
and not verify provided input when change IPv4.

This allows users passing fields that they only want to change instead
of passing all the fields. The old implementation would delete the
want-to-changed IPv4 address without creating it again, because it
expected all fields are provided.

This commit also moves the null entry check above to make it more easy
to read.

Tested with following inputs:
 + Full fields provided (address, netmask, and gateway)
 + Missing one, or two fields.
 + Empty input ie. {}.

Fixes #2270

Change-Id: I4a2dfd7165ba169971f5ce811a999d5cf96d98ad
Signed-off-by: Hy Che <hche@amperecomputing.com>