Ampere: Ethernet: Support InterfaceEnabled property.
the property is mapped by "Functional" property of OperationalStatus
interface in dbus. So the functions only work when the patch from
phosphor-networkd is merged.

InterfaceEnabled in this case means the interface is UP not RUNNING.
Therefore, if there are interfaces not receiving any packets and it
is UP then InterfaceEnabled will be "true".

Functions are:
+ retrieve InterfaceEnabled property and show it through GET request.
+ Change interface enabled through PATCH request (accepted values
  true/false).

Tested by sending GET/PATCH requests manually.

Implement #2399

Change-Id: I103e087723203a0782f9c67cf69ec3cde592d812
Signed-off-by: Hy Che <hche@amperecomputing.com>