Ampere: managers: Remove OEM property
It causes RedfishServiceValidator tool failed, because the redfish
standard expects a new schema for added OEM. Also, this property
is not mandatory by OCP profile, will add again when we have more
knowdledge about it.

Tested again with RedfishServiceValidator, and the error was gone since
there was nothing to check.

Fixes #2394

Change-Id: Icd1c6367fe4c5675cf38b55326638ec69705d2fa
Signed-off-by: Hy Che <hche@amperecomputing.com>