Ampere: Not allow destroy response after processing..
when there is "connection: close" in request's header field.

Because redfish in bmcweb using asynchronous methods, which call to
D-Bus and retrieve information, in each method handlers (eg: doGet,
doPatch, etc), which means there are threads that still work with
the resource when the main handler is finished. In case of the
request has "connection: close", the resources are immediately destroyed
after finish processing, this make bmcweb crashed since the asynch methods
are still accessing the destroyed resources.

The fix prevents destroying reponse in do_read() function, and only allow
setting is_reading flag to "false", the flag indicates there are no more
requests on the current connection. Therefore, later on do_write()
function, it can destroy the resource only after the response is finished,
without crashing bmcweb.

Tested again with requests contain "connection: close" in header. There
are no more crashes, and current connection is closed when finished.

Fixes #2526

Change-Id: Icaaaaeacc4c1ce5352ea462d0ae6efd440a61a53
Signed-off-by: Hy Che <hche@amperecomputing.com>