Ampere: Cleans up NetworkProtocol
Previously, the protocol's state is got from the service's state that
handles the protocol. It is true in some cases, but not in SSH, since
the service (dropbear.service) handles it only active in occasional
case, and we can still make a SSH connection.

The change gets the protocol's state from the socket's state instead.
This also allows implementing PATCH methods more easily.

Resolves #2901

Change-Id: I8565061405cf0f26cfb1bd3efdc230320b0dc195
Signed-off-by: Hy Che <hche@amperecomputing.com>